+++ components/script/Cargo.toml
@@ -70,6 +70,7 @@ libc = "0.2"
 log = "0.4"
 malloc_size_of = { path = "../malloc_size_of" }
 malloc_size_of_derive = "0.1"
+media = {path = "../media"}
 metrics = {path = "../metrics"}
 mitochondria = "1.1.2"
 mime = "0.3.13"
