+++ components/script/dom/window.rs
@@ -68,7 +68,6 @@ use crate::webdriver_handlers::jsval_to_webdriver;
 use app_units::Au;
 use base64;
 use bluetooth_traits::BluetoothRequest;
-use canvas_traits::media::WindowGLContext;
 use canvas_traits::webgl::WebGLChan;
 use crossbeam_channel::{unbounded, Sender, TryRecvError};
 use cssparser::{Parser, ParserInput, SourceLocation};
@@ -86,6 +85,7 @@ use js::jsval::JSVal;
 use js::jsval::UndefinedValue;
 use js::rust::wrappers::JS_DefineProperty;
 use js::rust::HandleValue;
+use media::WindowGLContext;
 use msg::constellation_msg::PipelineId;
 use net_traits::image_cache::{ImageCache, ImageResponder, ImageResponse};
 use net_traits::image_cache::{PendingImageId, PendingImageResponse};
