+++ components/script_traits/Cargo.toml
@@ -27,6 +27,7 @@ keyboard-types = "0.4.3"
 libc = "0.2"
 malloc_size_of = { path = "../malloc_size_of" }
 malloc_size_of_derive = "0.1"
+media = {path = "../media"}
 msg = {path = "../msg"}
 net_traits = {path = "../net_traits"}
 pixels = {path = "../pixels"}
