+++ components/servo/Cargo.toml
@@ -55,6 +55,7 @@ ipc-channel = "0.11"
 keyboard-types = "0.4"
 layout_thread = {path = "../layout_thread"}
 log = "0.4"
+media = {path = "../media"}
 msg = {path = "../msg"}
 net = {path = "../net"}
 net_traits = {path = "../net_traits"}
